pyproject.toml
[project]
name = "ytranslate"
version = "0.1.0"
description = "YouTube Video Translator and TTS"
requires-python = ">=3.14"
dependencies = [
    "streamlit>=1.5.0",
    "youtube-transcript-api==1.2.4",
    "openai>=1.0.0",
    "pydub>=0.25.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "audioop-lts",
    "pyaudioop",
    "yt-dlp>=2025.12.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
